+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED 
+### Added
+- Create file based in the type of analysis, number of elements and initial and final angular frequency.
+- AnalysisType in OperationRequestBase.
+- Finite Element Procedures - Klaus-Jurgen Bathe in bibliography.
+### Changed
+- Input.Parameter receives each parameter of Request.NewmarkMethodParameter.
+### Removed
+- AnalysisExplanation in OperationRequestBase.
+
 ## [1.1.0] - 2020-03-19
 ### Added
 - Programs in C to solution.